+++ templates/cjs/routes/users.js
@@ -2,6 +2,7 @@ const express = require('express');
 const UserModel = require('../models/UserModel');
 
 const router = express.Router();
+
 router.get('/', (req, res) => {
   res.render('users');
 });
@@ -15,4 +16,5 @@ router.get('/edit-personal', async (req, res) => {
   const user = await UserModel.getUser(req.user.id);
   res.render('edit-personal', {user: user.toJSON()});
 });
+
 module.exports = router;