+++ templates/esm/routes/users.js
@@ -2,6 +2,7 @@ import {Router} from 'express';
 import UserModel from '../models/UserModel.js';
 
 const router = Router();
+
 router.get('/', (req, res) => {
   res.render('users');
 });
@@ -15,4 +16,5 @@ router.get('/edit-personal', async (req, res) => {
   const user = await UserModel.getUser(req.user.id);
   res.render('edit-personal', {user: user.toJSON()});
 });
+
 export default router;